Fix compilation (for real this time)
