--- src/main.h
@@ -1,8 +1,0 @@
-#pragma once
-#ifndef CATA_SRC_MAIN_H
-#define CATA_SRC_MAIN_H
-
-[[ noreturn ]]
-void exit_handler( int status );
-
-#endif // CATA_SRC_MAIN_H
