+++ src/runtime_handlers.h
@@ -1,0 +1,8 @@
+#pragma once
+#ifndef CATA_SRC_RUNTIME_HANDLERS_H
+#define CATA_SRC_RUNTIME_HANDLERS_H
+
+[[ noreturn ]]
+void exit_handler( int status );
+
+#endif // CATA_SRC_RUNTIME_HANDLERS_H
